backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.4
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.3.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-yasg==1.21.7
email-validator==2.0.0.post2
fastapi==0.100.0
filelock==3.12.0
Flask==2.3.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.17.0
numpy==1.26.0
orjson==3.9.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
pipenv==2023.6.2
platformdirs==3.5.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.1
pydantic_core==2.1.2
pydivert==2.1.0
PyJWT==2.7.0
pyodbc==4.0.39
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
selenium==4.13.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.27.0
transbank-sdk==4.0.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.22.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
watchfiles==0.19.0
webdriver-manager==4.0.1
websockets==11.0.3
Werkzeug==2.3.6
wsproto==1.2.0
zstandard==0.21.0
